small fix + rename scollable -> scrollable
