+++ lib/src/screens/nano/nano_change_rep_page.dart
@@ -14,7 +14,7 @@ import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:cake_wallet/generated/i18n.dart';
 import 'package:cake_wallet/src/widgets/primary_button.dart';
 import 'package:cake_wallet/src/screens/base_page.dart';
-import 'package:cake_wallet/src/widgets/scollable_with_bottom_section.dart';
+import 'package:cake_wallet/src/widgets/scrollable_with_bottom_section.dart';
 
 class NanoChangeRepPage extends BasePage {
   NanoChangeRepPage({required SettingsStore settingsStore, required WalletBase wallet})
