+++ lib/src/screens/new_wallet/advanced_privacy_settings_page.dart
@@ -16,7 +16,7 @@ import 'package:flutter/material.dart';
 import 'package:cake_wallet/generated/i18n.dart';
 import 'package:cake_wallet/src/screens/base_page.dart';
 import 'package:cake_wallet/src/widgets/primary_button.dart';
-import 'package:cake_wallet/src/widgets/scollable_with_bottom_section.dart';
+import 'package:cake_wallet/src/widgets/scrollable_with_bottom_section.dart';
 
 class AdvancedPrivacySettingsPage extends BasePage {
   AdvancedPrivacySettingsPage(
