+++ lib/src/screens/yat/widgets/second_introduction.dart
@@ -1,7 +1,7 @@
 import 'package:cake_wallet/src/screens/yat/widgets/yat_bar.dart';
 import 'package:cake_wallet/src/screens/yat/widgets/yat_page_indicator.dart';
 import 'package:cake_wallet/src/widgets/primary_button.dart';
-import 'package:cake_wallet/src/widgets/scollable_with_bottom_section.dart';
+import 'package:cake_wallet/src/widgets/scrollable_with_bottom_section.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:cake_wallet/palette.dart';
